data info refactor
